Feature/no ref authentication

+++ src/posts/comments/comments.service.ts
@@ -1,5 +1,4 @@
-import { Injectable } from '@nestjs/common';
-import { UpdateCommentDto } from './dto/update-comment.dto';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { Prisma, Comment } from 'generated/prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 
@@ -11,11 +10,15 @@ export interface ICommentSearchParams {
 export class CommentsService {
   constructor(private prisma: PrismaService) { }
 
-  async createPostComment(postId: number, commentInput: Prisma.CommentCreateInput): Promise<void> {
-    this.prisma.post.update({
-      where: { id: postId },
-      data: { comments: { create: commentInput } }
-    })
+  async createPostComment(postId: number, userId: number, commentInput: Prisma.CommentCreateInput): Promise<void> {
+    try {
+      await this.prisma.post.update({
+        where: { id: postId, userId },
+        data: { comments: { create: commentInput } }
+      })
+    } catch {
+      throw new BadRequestException()
+    }
   }
 
   async findPostComments(postId: number): Promise<Comment[]> {
@@ -27,14 +30,14 @@ export class CommentsService {
     return this.prisma.comment.findUniqueOrThrow({ where: { id } })
   }
 
-  async update(id: number, data: Prisma.CommentUpdateInput): Promise<Comment> {
-    return this.prisma.comment.update({
-      where: { id },
+  async update(id: number, userId: number, data: Prisma.CommentUpdateInput): Promise<Comment> {
+    return await this.prisma.comment.update({
+      where: { id, userId },
       data
     })
   }
 
-  async remove(id: number): Promise<void> {
-    this.prisma.comment.delete({ where: { id } })
+  async remove(id: number, userId: number): Promise<void> {
+    this.prisma.comment.delete({ where: { id, userId } })
   }
 }
